+++ HISTORY.rst
@@ -2,6 +2,12 @@
 History
 =======
 
+0.2.0 (2022-03-27)
+------------------
+
+* Fixed an error on forecast.price() when the API doesn't return all keys
+* Added initial work on consumption stats
+
 0.1.1 (2022-03-26)
 ------------------